+++ tests/input_value.rs
@@ -1,6 +1,6 @@
 use async_graphql::*;
 
-#[async_std::test]
+#[tokio::test]
 pub async fn test_input_value_custom_error() {
     struct Query;