+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.14] - 2021-03-14
+
+- Add `DataLoader::loader` method. [#441](https://github.com/async-graphql/async-graphql/issues/441)
+- Fix the validation does not work on some inline fragments.
+
 ## [2.5.13] - 2021-03-09
 
 - Support generics in Subscription types. [#438](https://github.com/async-graphql/async-graphql/pull/438)